+++ pricing/src/epilot/models/shared/cartdto.py
@@ -3,6 +3,7 @@
 from __future__ import annotations
 import dataclasses
 from ..shared import customer as shared_customer
+from ..shared import metadata as shared_metadata
 from ..shared import ordersource as shared_ordersource
 from ..shared import orderstatus as shared_orderstatus
 from ..shared import paymentmethod as shared_paymentmethod
@@ -27,7 +28,7 @@ class CartDto:
     files: Optional[list[str]] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('files'), 'exclude': lambda f: f is None }})
     r"""An array of file IDs, already upload into the File API, that are related with this cart"""
     journey_data: Optional[dict[str, Any]] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('journey_data'), 'exclude': lambda f: f is None }})
-    metadata: Optional[list[Union[MetaData1]]] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('metadata'), 'exclude': lambda f: f is None }})
+    metadata: Optional[list[Union[shared_metadata.MetaData1]]] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('metadata'), 'exclude': lambda f: f is None }})
     r"""A set of key-value pairs used to store meta data information about an entity."""
     payment_method: Optional[shared_paymentmethod.PaymentMethod] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('payment_method'), 'exclude': lambda f: f is None }})
     r"""A PaymentMethod represent your customer's payment instruments."""